+++ CHANGELOG.md
@@ -342,9 +342,11 @@ Released 2024-01-11
 ### Version 0.6.5
 Released 2024-01-25
 
-- Extends Response object to be more useful when making lambda request
-- Migrates away from datetime.utwnow() to datetime.now(UTC)
-- moves get_response from exceptions classes to Response class from_exception
+- Moves `get_response` from exceptions classes to Response class as `from_exception`
+- Deprecates `is_ok` method in favour to `ok` property
+- Extends Response object to be more useful when making lambda request by adding `json` method
+- Migrates away from `datetime.utwnow()` to `datetime.now(UTC)`
+
 
 ### Version 0.7.0
 Release ETA 2024-02-31 ;)